backend/requirements.txt
fastapi
uvicorn
docling
python-dotenv
openai
langchain-openai
langchain-qdrant
qdrant-client
langchain
python-multipart